# refactor: optimize core dump and config checks