<!--
Please make sure you've read and understood our contributing guidelines;
https://github.com/sonic-net/SONiC/blob/gh-pages/CONTRIBUTING.md

Please provide following information to help code review process a bit easier:
-->
### Description of PR
<!--
- Please include a summary of the change and which issue is fixed.
- Please also include relevant motivation and context. Where should reviewer start? background context?
- List any dependencies that are required for this change.
-->
Optimize `core_dump_and_config_check()` by Python multithreading to reduce the running time.

Summary:
Fixes # (issue) Microsoft ADO 30849826

### Type of change

<!--
- Fill x for your type of change.
- e.g.
- [x] Bug fix
-->

- [ ] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] New Test case
    - [ ] Skipped for non-supported platforms
- [x] Test case improvement


### Back port request
- [ ] 202012
- [ ] 202205
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [x] 202411

### Approach
#### What is the motivation for this PR?
In `core_dump_and_config_check()`, we will run many commands on all the DUTs in sequence, which will take a long time to finish on multi-DUT devices, such as T2 devices. For a standard Nightly, we will basically run `core_dump_and_config_check()` for every non-skipped test module, which will accumulate to a very long time. Therefore, we wanted to reduce the running time by parallelizing the commands using Python multithreading.

#### How did you do it?

#### How did you verify/test it?
I ran the updated code and can confirm it's working well. After the optimization, the running time of `core_dump_and_config_check()` has the following improvements:
- For dual ToR topo, the setup running time will be decreased from 5s to 3s, the teardown running time will be decreased from 8s to 4s (Elastictest [link](https://elastictest.org/scheduler/testplan/67d7543e607a6896f60ddd9c?testcase=bgp%2Ftest_bgp_fact.py&type=log) after the change)
- For T2 topo, the setup running time will be decreased from 45s to 20s, the teardown running time will be decreased from 47s to 24s (Elastictest [link](https://elastictest.org/scheduler/testplan/67d79e1606bd43461b9eecb2?testcase=pc%2Ftest_po_cleanup.py&type=console) after the change)


#### Any platform specific information?

#### Supported testbed topology if it's a new test case?

### Documentation
<!--
(If it's a new feature, new test case)
Did you update documentation/Wiki relevant to your implementation?
Link to the wiki page?
-->